make-public: factor out FuzzyMatcher
